test(cli): the integration tests follow the redesigned screens
`session show`, `session search`, `config set` and `version`/`update` all
render differently now, so the assertions that pinned their old box art
move onto the new sections. `version` gains an assertion for the block it
exists to print: the invoked name, the index path, and every parser
version. `update` gains one for what it must never do -- print an install
line when there is nothing to install.

+++ apps/cli/src/app.rs
@@ -2453,7 +2453,7 @@ pub(crate) fn inspect_view(
     let strongest = agentworth_outcomes::highest_outcome(&outcomes);
     let outcome = strongest.map(|s| {
         let rung = outcome_rung(Some(s.kind));
-        let supporting: Vec<(String, f64)> = outcomes
+        let supporting: Vec<(String, f32)> = outcomes
             .iter()
             .filter(|o| o.kind == s.kind)
             .map(|o| (o.summary.clone(), o.confidence))
